test: RecommendedArtworks
