+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -59,148 +59,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("RecommendedArtworks", () => {
-    it("returns lists of artworksConnection", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(id: "home-view-section-recommended-artworks") {
-              __typename
-              component {
-                title
-                behaviors {
-                  viewAll {
-                    buttonText
-                    href
-                    ownerType
-                  }
-                }
-              }
-              ownerType
-
-              ... on HomeViewSectionArtworks {
-                artworksConnection(first: 2) {
-                  edges {
-                    node {
-                      id
-                      title
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const vortexResponse = {
-        data: {
-          artworkRecommendations: {
-            edges: [
-              {
-                node: {
-                  artworkId: "608a7417bdfbd1a789ba092a",
-                  score: 3.422242962512335,
-                },
-              },
-              {
-                node: {
-                  artworkId: "308a7416bdfbd1a789ba0911",
-                  score: 3.2225049587839654,
-                },
-              },
-              {
-                node: {
-                  artworkId: "208a7416bdfbd1a789ba0911",
-                  score: 4.2225049587839654,
-                },
-              },
-              {
-                node: {
-                  artworkId: "108a7416bdfbd1a789ba0911",
-                  score: 5.2225049587839654,
-                },
-              },
-            ],
-            totalCount: 4,
-          },
-        },
-      }
-
-      const vortexGraphQLAuthenticatedLoader = jest.fn(() => async () =>
-        vortexResponse
-      )
-
-      const artworksResponse = [
-        {
-          _id: "608a7417bdfbd1a789ba092a",
-          id: "gerhard-richter-abendstimmung-evening-calm-2",
-          slug: "gerhard-richter-abendstimmung-evening-calm-2",
-        },
-        {
-          _id: "308a7416bdfbd1a789ba0911",
-          id: "pablo-picasso-deux-femmes-nues-dans-un-arbre-2",
-          slug: "pablo-picasso-deux-femmes-nues-dans-un-arbre-2",
-        },
-      ]
-
-      const artworksLoader = jest.fn(async () => artworksResponse)
-
-      const context: any = {
-        accessToken: "424242",
-        artworksLoader,
-        userID: "vortex-user-id",
-        authenticatedLoaders: {
-          vortexGraphqlLoader: vortexGraphQLAuthenticatedLoader,
-        },
-        unauthenticatedLoaders: {
-          vortexGraphqlLoader: null,
-        },
-      }
-
-      const response = await runQuery(query, context)
-
-      expect(artworksLoader).toHaveBeenCalledWith({
-        ids: ["608a7417bdfbd1a789ba092a", "308a7416bdfbd1a789ba0911"],
-      })
-
-      expect(response.homeView).toMatchInlineSnapshot(`
-        Object {
-          "section": Object {
-            "__typename": "HomeViewSectionArtworks",
-            "artworksConnection": Object {
-              "edges": Array [
-                Object {
-                  "node": Object {
-                    "id": "QXJ0d29yazo2MDhhNzQxN2JkZmJkMWE3ODliYTA5MmE=",
-                    "title": "Untitled",
-                  },
-                },
-                Object {
-                  "node": Object {
-                    "id": "QXJ0d29yazozMDhhNzQxNmJkZmJkMWE3ODliYTA5MTE=",
-                    "title": "Untitled",
-                  },
-                },
-              ],
-            },
-            "component": Object {
-              "behaviors": Object {
-                "viewAll": Object {
-                  "buttonText": "Browse All Artworks",
-                  "href": null,
-                  "ownerType": null,
-                },
-              },
-              "title": "Artwork Recommendations",
-            },
-            "ownerType": "artworkRecommendations",
-          },
-        }
-      `)
-    })
-  })
-
   describe("HomeViewSectionViewingRooms", () => {
     it("returns correct data", async () => {
       const query = gql`
